requirements.txt
mcp[cli]
supabase
python-dotenv
pytest
pytest-mock 